QuickOpen optimization, Other option to drop boxes
 - Оптимизация функции "Открыть быстро"
 - Добавлена возможность ввода не предусмотренных значений в дроп-боксы в функциях с дочерними интерфейсами, для этого
        нужно выбрать пункт "Другие" в любом из дроп-боксов, в котором есть такой пункт
 - Добавлена "скрытая" опция в файл настроек для вкл\откл режим отладки

+++ source/reapers/other_prg.py
@@ -1,5 +1,6 @@
 
 from random import randint
+from time import sleep
 from icecream import ic
 from subprocess import Popen, PIPE
 
@@ -19,6 +20,7 @@ class OtherProg(Reaper):
         self.first_arg = ''
         self.second_arg = ''
         self.splitter = ' '
+        self.real_pb = True
 
     @file_reaper
     def run(self):
@@ -32,31 +34,44 @@ class OtherProg(Reaper):
         ic(self.file_name)
         ic(f'"{self.path_to_root}data/{self.program_name}" {self.first_arg} "{self.file_name}" {self.second_arg}')
 
-        while True:
-            out = prg.stdout.readline().strip()
-            o = out.split(self.splitter)
-            percent = 0
-            ic(out)
+        if self.real_pb:
 
-            try:
+            while True:
+                out = prg.stdout.readline().strip()
+                o = out.split(self.splitter)
+                percent = 0
+                ic(out)
 
-                if self.percent_type == 'int':
-                    percent = int(o[self.percent_index])
-                elif self.percent_type == 'l_first':
-                    first, second = o[self.percent_index].split(self.percent_del)
-                    percent = int(100 / int(second) * int(first))
-                elif self.percent_type == 'b_first':
-                    first, second = o[self.percent_index].split(self.percent_del)
-                    percent = int(100 / int(first) * int(second))
-                else:
-                    percent += randint(1, 5) if percent <= 99 else 99
+                try:
 
-                self.update_signal.emit(percent, '', f'{localize.saving} - {o[self.name_index]}...', False)
-            except (ValueError, IndexError):
-                pass
+                    if self.percent_type == 'int':
+                        percent = int(o[self.percent_index])
+                    elif self.percent_type == 'l_first':
+                        first, second = o[self.percent_index].split(self.percent_del)
+                        percent = int(100 / int(second) * int(first))
+                    elif self.percent_type == 'b_first':
+                        first, second = o[self.percent_index].split(self.percent_del)
+                        percent = int(100 / int(first) * int(second))
+                    else:
+                        percent += randint(1, 5) if percent <= 99 else 99
 
-            if not out:
-                prg.kill()
-                break
+                    self.update_signal.emit(percent, '', f'{localize.saving} - {o[self.name_index]}...', False)
+                except (ValueError, IndexError):
+                    pass
+
+                if not out:
+                    prg.kill()
+                    break
+
+        elif self.real_pb is None:
+            pass
+
+        else:
+            a = 0
+
+            while prg.poll() is None:
+                a += randint(1, 5) if a <= 99 else 99
+                self.update_signal.emit(a, '', f'{localize.unpacking} - {self.file_name}...', False)
+                sleep(randint(1, 5))
 
         self.update_signal.emit(100, '', localize.done, True)
